yaml: forbid includes from parent directories
Only allow including files that are in the same directory than the
including file (or in subdirectories).

Yes, this is very restrictive, but it provides a strong security w.r.t
repacking:
* we ensure we only try to create files in the output directory, and not
  write over files all over the filesystem.
* we ensure the data can be repacked in a single self-contained
  directory, which is a guarantee that we want (ie, that repacking does
  not leave dependencies on files outside of the output directory, which
  could invalidate the repacked data).

It remains to be seen if this is satisfactory, or if we will want some
more freedom on inclusion. In particular, we could relax the constraint
so that includes from parent are allowed, but only if still contained in
the "root" output directory. See for example the include test: it would
still be valid with this rule.

Change-Id: I9a024c7844a463949942f10cf62677d68597c713

rip-it: 80c8238c026efeab8cc80a5ecb546a5bed0d8ba3