yaml: update API to use 'const char *' for path rather than lstr_t
Use C string pointers for all public APIs that expect a path. This is
better as it explicit the fact that the string is C-style, while a
lstr_t can be constructed with a missing trailing 0.

The code of the attach_file API is reworked a bit to use path_simplify
instead of canonical: we do not care about having a canonical path, we
only care about resolving all path movements.

Change-Id: I972bc55b314522b9b39619d1bcb112f92d42947e

rip-it: cef72191a33951b8126be158a1d8ef6de43f41f4