iop-yaml: handle unpacking integers into double fields
This conversion is trivial to handle. It is especially necessary as a
double that is equal to an integer will be packed as an integer:

 double_value: 3

and not "3.". This could be modified in the future, but at the very
least we have to handle this conversion.

Change-Id: Ib5467bf2c86f49515a33977ffaf8bc2942257b15

rip-it: cf00d8995769bd448d43b1e914853e626a53c42a