iop-yaml: allow pack/unpack of private fields by default
Private fields are not about being unknown to the public, but rather
about being undocumented/hidden, but must still be accessible and
properly packed.

The behavior is thus changed to handle private defaults properly by
default, and let the user be able to forbid their use with a flag.

Change-Id: I11865814bafbd07c87e9526173edf390a9705d77

rip-it: 3ea314946b07421509c32ec23273480fa6ef2f21