yaml: start handling overrides in presentation
Handling of overrides in presentation data is far from trivial. It
changes several assumptions that were working before, for example:
 * The AST is never changed, so repacked included files can have their
   AST compared to find out if they are different. This is now broken.
 * Only presentation data needs serializing, not AST values. This is no
   longer true.

As we want to end up with a parse AST that has no notion of inclusion,
overrides etc, this AST contains overriding values. This means we have
to save the overridden values in the presentation data, so that they
can be reused when repacking.

The design is to save for an override a list of the values that it
replaced. Those values are stored with their path from the override's
root node. Then, when repacking:
* if a node's path matches an override path, the original value is
  retrieved and packed. The actual value is stored in the override data,
  in place of the original value.
* a full yaml_data_t object is rebuilt from the (path, values) of an
  override object (remember, those values are the ones that were in the
  AST, not the original values anymore).
* the data is packed just after the node that was overridden.

This may seems a bit convoluted, but it will be quite robust, especially
given the situations that can occur, and that will be tested in
following commits:
* override of an AST that when through several includes
* override of another override.
* ...

A real description of how presentation data works (a bit like in this
commit, but much more detailed) will be added to the documentation, once
the feature is stable.

Change-Id: I77e0ed8502662b2beee71df7d2ee684932d8be89

rip-it: f3908d573057c6ca6de46b9b7105aadd64d82f40