yaml: prevent loss of data when packing in flow mode
Flow mode is a particular case for YAML, where some features are not
available, mainly tags.
However, as presentation can be separated from the data, it is possible
to end up when repacking with a node that was packed into flow mode but
that now contains tags.

To handle this, a look-ahead is done to ensure we can pack in flow mode.
This isn't ideal, but the impact should be quite low, given that flow
data are usually very small.

Change-Id: Ib4f9de3145e7f4157a001cbdb429c6ad500a3c19

rip-it: a97a8c715d7bc802d6cf6bf0d0b2f62c5da132fc