yaml: handle presentation of override data
Save the presentation of the override data, so that it can be reapplied
when repacking the override.
As the document presentation is created with an empty path as reference,
the current_path of the env must be updated as well, before packing the
override data.

Change-Id: Ia918cfc8c954ae42c8d4b770539470fe9edbe6aa

rip-it: 05a04d4be2ac07eb1aad6a6bb547999fd32518c8