yaml: properly handle overrides with other overrides or includes
This is where it is starting to get complex.

First issue, an override can modify an AST that was included from
another file, itself included from another file, etc. Therefore, the
"path" that we use to match a local data with the environment override
may not match, as the local path may be much smaller (see test: the root
file modify a data through the path g.c, but the data comes from a file
where the path is .c).
To handle this, we now keep track of an absolute path. This allows
matching an override that was done in any point, through any includes.
This is still a bit rough, as we do a lot of allocations to update both
paths. This should be improved later on.

Second issue, we may stack up overrides. A local data may have been
overridden by any of those overrides, and even better, an override may
have remodified a data that was modified by an earlier override.
To handle this, we keep a stack of overrides. To match a local data,
we must go through the stack in LIFO order, so that the local data is
swapped with the latest override. This override itself, when packed,
will also go through the active overrides, to itself swap its values
with other overrides.
This mechanism properly bubbles up the actual value from the AST to the
latest override that modified it.

Change-Id: Idb8785d43dd5d1d3e4aa4dc392f52253422cc1e3

rip-it: 19fa2fd4671507fefd5e26a47276fee37c2990ee