yaml: simplify handling of path when packing
We need both the absolute path (from the root document) and the current
path (from the current file's root) when packing. Instead of keeping two
separate buffers and doing lstr_fmt in a lot of places, simplify this
by:
* keeping a single buffer: the absolute path
* keeping the index where the "current" portion of the path starts.

Now, absolute path can be used as is, and the current path is easily
computed by creating a custom lstr that starts from the given position.

Change-Id: I157fed2a94a379d084c02ff3f4b288de86fbf4fa

rip-it: c7982d85890ebbde244c5dab5d03b772e83fc6ae