yaml: handle overrides with YAML_PACK_NO_SUBFILES flag
This flag activates a behavior a bit weird compared to the rest. Instead
of either packing the AST into a single file, or recreating the
subfiles, it repacks the root file as it should look like, but do not
recreate the subfiles.

To handle overrides, we must do as if we pack the subfiles, so that we
iterate on the inner AST and properly rebuild the overrides.
The code is a bit modified to factorize helpers between the classic case
& the NO_SUBFILES one.

Change-Id: I15b4a019109c0655dff0df06bca01fb0278fea83

rip-it: 0798f46b974dbd42a7b1f0c17a526a5506415628