yamlfmt: add dso argument to validate IOP YAML documents
Use the dso to get the IOP struct described by the top-level tag, and
validate the yaml data by parsing the IOP object.

This requires a new iop-yaml API, to parse a yaml data (and not a
string) into an IOP object.

Change-Id: Ie55d0a11484476cd586bcc3cc8222c5f0057bf7d

rip-it: 45a7418eaf25a43e8b16956570db1851ec16f871