yaml: handle templated variables
Detect use of "$<name>" in a document, and save the location of those
uses. When a file is included, those variables can be specified, and
replaced in the AST.

This is still very early in the implementation, a lot will change in the
following commits.

Change-Id: Id93844b8c6dfca943bbd7cd379f57e64b7dc3df9

rip-it: fa0c665f7e06947ce46d3d8cfad4844960497c3e