yaml: separate variables handling from override
Split the handling of variables setting from the override. The variables
settings has an object syntax that may not match the included AST (as
the fixme in the test indicates), while the override must have the same
type.

Change-Id: I513de0134833cd1088bed28245bb2428de65c728

rip-it: 9482d3bc12d577dc63ef6f813be9f9e58c7064f1