yaml: reject documents with unbound variables
In the majority of situations, we want the parsed AST to be complete,
without remaining variables. But in some others, we want to keep
variables unbound, for example when we are manipulating the YAML data
without interpretation.

This was the default behavior, but is now behind a parsing flag.

Change-Id: I28d3a0bad0cce6cd4c37ee27819b582f4c19158e

rip-it: 3c4d3a108c5c539343dffa19338df81b360c4d46