yaml: start handling variables in strings
We want to be able to handle variables that are used inside a string,
so that we are able to do things like "$host:$ip".

The implementation is started on only quoted strings, and only for a
single variable. The rest will come in the next commits.

Change-Id: I14c6e4536755a29905b2b6f38be6bea1cf738c0e

rip-it: fc8529d94829831957ea06f70bde18d9c64e712a