yaml: handle variables in unquoted strings, deduplicate them
Variables should be usable in unquoted strings as well. We need to
distinguish the case where the whole string is the variable (which means
it can be replaced by any AST), or a subpart (it can only be replaced by
a string).

Change-Id: I1e843bbbf32f885c6fd9801993692d0462359488

rip-it: 853d455f52915ea12683de81c5c0a31ecdf2880e