yaml: test using the same variable name multiple times
Make sure that if several variables context using the same variable name
are active at the same time, they are properly handled and the values
properly repacked.
This test showed a bug in the order of the vars repack: this is fixed as
well.

Change-Id: I6edba4156b243d91376d2cbda0b09820c2728256

rip-it: 16893ca143e496bc4bcede0e93b25757338df7f5