yaml: handle repacking for strings containing multiple variables
The previous behavior is to parse a templated string, to find out
the value of the variable:

template: "var is <$var>"
value: "var is <dada>"

=> value for var is data. So when repacking the override, we can pack:

 $var: dada

This works ok as long as the template does not contain multiple
variables. In that case, it becomes at best really hard to deduce
variable values (at least without a regex library), and in some cases
ambiguous:

template: "$foo $bar"
value: "a b c d e"

foo = ?, bar = ?

To still allow using multiple variables in strings on parsing, and
properly repack them, we need another way to get the variable values
back.

The design is now to store in the presentation data the value for
variables used in strings (we need to store this value, it is the only
way to get it and properly pack the variable settings).
Then, when packing a template strings, we apply those original values
to the template, and check if the result matches the AST. That should
be the case almost all the time, and allows repacking those values
properly.

However, as soon as a mismatch is found, we do not try anything at all,
and will pack without the template, and lose the template.
Those kinds of conflicts will be tested more thoroughly in the next
commit.

Change-Id: I856b002c34c9f0b3f8505e0cc128585980c0faca

rip-it: 7c03d9b95e6dfdc76993428b42434602bc079225