yaml: test modification handling in several cases
Test how changes in the AST is handled if the AST was introduced through
variables.
For raw templates, this is trivial, and always works.
For a string using a single variable, this works as well, as long as the
full string still matches the template.
For a string using multiple variables, no changes are handled.

What is missing is handling the conflict of a shared variable with
different values.

Change-Id: Ibe61c8856089464adb91266301a68deaffaa393d

rip-it: a80bba7a458d721c51b0edd86339613cb6afd358