yaml: allow bracketing variable expressions
We want to be able to specify variable names with brackets, to prevent
collision with surrounding strings: "${a}a" for example.

To do so, a new "parse_variable_name" routine is added, and the rest
of the code is adapted to handle both cases.

Change-Id: I50f15f28be84839a1401a333caf76b76fed4a1d1

rip-it: 2120814e0ca5833bfbbcfc6bc55d8e8618ffbd76