yaml: start handling variable resolution conflicts
Variables are often re-used in multiple places, which leads to issues
if, when repacking, those values are not equal.

For example:

 - $var
 - $var
 - $var

With the AST:

 - 1
 - 2
 - 2

In those cases, we have two choices:

1/ lose the templating, and pack the raw values:

- $var
- 2
- 2

This can lead to weird results if the value we choose is the least used,
as shown in the example. In addition, we lose semantic meaning and
remove variables that can be meaningful, and was purposefully used by
the user.

2/ add new variables:

- $var
- ${var~1}
- ${var~1}

This makes the code more complex, but keeps the use of variables, and
is more meaningful to the user when resolving the conflicts. The user
can easily see that those were previously the same variables, but some
changes caused the conflict, and he come resolve those conflicts more
easily.

This second solution was picked. When a different data value is found,
an available new name is searched, and used for this new data.

The complexity lies in the rewriting of existing templates to use the
new name. This will be done in the next commits.

Change-Id: I08edbb54bf15b2a462e26cd2cd5a5f67332bb0df

rip-it: 6656619e792a9a6656cda68f2d034d84e1e5c283