yaml: add new flag on packing to allow unbound variables
It is possible to parse a YAML data and allow unbound variables. In
order to repack this data the same way, unbound variables should be
allowed on repack as well.

This is implemented and tested, and used for yamlfmt.

Change-Id: I6c3890ec0f82470b704f4e6ca0b6863405c7fb1e

rip-it: b36a02dae5be2e441008b5c98ec5e60796d7f01b