yaml: handle var substitution introducing dollar characters
If a var value contains '$', it must be handled in the var bitmap, to
not consider it as a variable.
This forces the bitmap to be nonnull in functions using it.
This isn't very pretty though :/ I'd like to find a better way to handle
this.

Change-Id: I8921671c3205ea39f6a98bcf4cec7be5bcd14489

rip-it: e7f43df35fdfaa827bf7b5f0700719f760872e6b