yaml: change variable syntax to $()
Previously, it was possible to write variables in two ways:

- $foo, if the name was pure alnum
- ${foo}, for more complex names, or to properly seperate the variable
  from the rest of the string.

This is however buggy, notably in the case of partial substitutions.
For example:

"$foo$bar", set $bar to "a" => "$fooa". When the string is then parsed
to replace $foo, it won't work, because the variable is now "$fooa".

To fix this, remove the ambiguous syntax, and force a stricter one.
Instead of ${}, I went for $(), for reasons.

This simplifies the code slightly, but makes the templates more
readable. In addition, the complex code handling all the escaped
variables has now even less chance of being triggered, as the sequence
to escape is '$(', which should be really rare. So all this complex code
should stay relatively cold.

Change-Id: I219b8840f03aec504e73814abc9662656b45080d

rip-it: 2ba059ea79ba4906826a92f139d2aea7e02ab954