yaml: test span of empty null scalar in errors
An empty line of EOF means a null scalar. However, this makes the span
of the data not really existent. Test in a case where we will print an
error using this span, eg with the variable settings.
The error message is not ideal, but should be enough for now.

Change-Id: I7c21db7604dbbbd05d43d89f1f7552ac506025ba

rip-it: f55f55337b986105709094b256416167c217bbb4