yaml: hide ~ after include in default cases
When converting IOP json includes to YAML presentation, we only generate
the "included" node, and do not fill the include presentation. This
leads to generating:
 !include:path ~
which is not very pretty.
Instead of fixing that in the json include -> yaml presentation
converter, do it by default in the packer:
 * this prevents growing the document presentation for no good reason
 * code generating presentation should not have to bother with this kind
   of adjustments.

Change-Id: Ib3bf647ee49a0587bbf8f9560af8f30f0d958735

rip-it: c29c7fac4eafaa8de5cfc5dcff89a88f92c0738c