iopc: fix dependency issues with classes ancestors
When generating the C struct associated to a class, all the attributes of
all the ancestors are pasted into the struct attributes, therefore, the
class doesn't only depend on the parent package but also on all the
ancestors packages.

Adding this direct dependency fixes some cases of missing IOP package
builds that could trigger surprising crashes or memory corruptions because
of the casts we could do between incompatible classes.

We should do the same for all structs inlined in other structs but this is
less problematic because the worst case can only trigger non optimal
struct packing.

Progress: 50%
Change-Id: I43c56261d3ceb331e94e4318a3fae85287a34306

rip-it: 5b77a3747af2716ac66979e005e43bb0de650f6a uprooted