zbehave: fix zbehave formatter for behave 1.2.6.
In the result method, step_result.status is now an enum and not a string
anymore.
We need to convert it back to a string in order to be compatible.

The feature method is not called when the feature is skipped and
show_skipped is False.
We must force show_skipped to True when using the ZFormatter in order to
capture the skipped features.

behave.formatter.formatters.register is deprecated since behave 1.2.5,
we should use behave.formatter._registry.register instead.

Change-Id: I8f9aec7cdc729c6c564b678b6786c9626350f304

rip-it: 5a5179c05c1bdc810a3a5d613f78cf2147fa73a3