iopy: support multiple upgrades for the same IOP type.
Before, a TypeError was raised when trying to upgrade multiple times the
same IOP type without `force_replace`. If `force_replace`, we used to
replace the previous upgrade with the new one.

The issue is that there are legitimate cases where you want to upgrade
an IOP type multiple times. For example, you can have a public upgrade
for release, and a private upgrade for tests to extend it.

This behaviour is also more consistent with the old metaclass upgrade
mechanism that supports multiple upgrades.

`force_replace` is now used to replace all the previous upgrades of an
IOP type. It is actually the same behaviour as before, since we used to
raise a TypeError if `force_replace` was not set in case of multiple
upgrades.

Change-Id: I0805591e3b1c656251ff9a3ed6093ba7ce20b65e

rip-it: 0cb6d11e0f60e474ef373efce4d894ed8d34a306