yaml: fix merge key repacking with value modified more than twice
As soon as the value for a key is found, we must break instead of going
on. Otherwise, values will be shuffled, when the one we want to pack is
the earliest we can find (the earliest it has been modified).

Change-Id: Ia755902383b86a1e004377bdfee9c62e832a2fb4

rip-it: f27fa6689df582469e10a142e57cefb047a44d57