yaml: add test to showcase issue with merge keys and overrides
Overrides do not work well with merge keys, due to complications on the
path (the AST path and the path in the merge key are different).
A big rework on how overrides are packed would be needed to fix this.

Add a test to at least show the issue and detect changes in how it is
handled.

Change-Id: I2a7de7a5ae62641963d3811a96d6f1c8c3c19c9e

rip-it: 3fa26ffee5ab7d09e13db1e70fd72be28bb81140