waf: Fix tasks rebuild when farchc/iopc/pxcc changes.
In Fc2c/Iop2c/Pxc2Pxd tasks, farchc/iopc/pxcc were added in the scan
method so that the task is re-executed when farchc/iopc/pxcc changes
(which is necessary because the output files can change when it happens,
even when the input ones don't).

It works fine, except in the case farchc/iopc/pxcc are moved between two
builds, because in that case the binaries at the previous location do not
change at the second build. This is what happens between branches 2019.3
and master (future 2020.04) because of the lib-common restructuration
series, and causes compilation errors when switching branches.

To fix that, simply add farchc/iopc/pxcc in the inputs of the
Fc2c/Iop2c/Pxc2Pxd tasks instead of using the scan method. Inputs are
explicit dependencies, so are not remembered from previous builds like
implicit dependencies of the scan method.

At first, it was not done like that because I considered having the
compiler in the inputs of a task was not natural, but this is the only way
I see to fix this bug.

Change-Id: Ia89158d2f58adee5b37a034f6147e92f9281c9df

rip-it: 1733b2d913aa7986cd8a52a09f02613e9c958e8b