yaml: always add tag on top level object
When packing an IOP class, always include the tag. This explicits the
type of the document, and prevents issues when repacking an object,
where it looks like we pack an object as its real class and not its base
class, bypassing the tag.

Change-Id: I38e97a505480eaddfe19a7b46583aeb96d8e5e60

rip-it: 957a81f6787c5f1d2fd5f8cf384c80db980791a4