zchk-iop: fix nr_47521 test
The test is checking that an invalid stream unpacked as a binary class
is properly handled. This does not give the right to the caller to give
undefined values however, as c_ptr was. So properly set this value. It
was not read in the test, but the next commit will read this value.

Also, remove a duplication of this test.

Change-Id: I98d9612de7730c8e019339c5631192cfdf87435b

rip-it: d563ccb0c50a505158f1f17a52edcb428fe525b6