build-version: use a tag name diffent from the product
When trying to determine the version of the product currently built, it
is now possible to look for a tag whose name is different from the
product. It can be useful if several products are "grouped" under a
tag called after a more generic name.

Change-Id: I8b49404732973b2581fd26c34ee0ef5db5f6ad29

rip-it: 5e4d591cacd588a8c9e435a52f384c538b840b54