core/compat: Provide a way to include system stdatomic.h.
In the next patch, I'll need to use the "_Atomic double" type, which does
not compile when using the compat stdatomic.h header.

The purpose of this patch is to provide a way of including the system
stdatomic.h in the rare cases where it is needed.

To do that, rename the compat herder to compat-stdatomic.h, and include
stdatomic.h (so the system one) if INCLUDE_SYS_STDATOMIC is defined.

Change-Id: Id2f2d1e98f4e03edd0346efedb42dec7adf4745a

rip-it: 3bbf48c2a8905481f1740da7b4045ff1c097b1f6