stdatomic: always use C11 stdatomic.h if available
Instead of using the shim in all situations (except for a bench exec),
include the standard header if possible, ie if gcc > 4.9 or clang is
used.

The shim is still used in C++ mode, because C11 is not C++ compatible.

Change-Id: I42fef9457872baf8bc4b61f21cf356fb6683526b

rip-it: 6ad7ed87361799c4dff656df1e13ea7366f7f2c0