core: define useful atomics typedef
Add typedef for all fixed sized integers atomics, unfortunately not
defined in the standard. Also, move the typedef for ssize_t there as
well, as it is not defined in stdatomic.h, so having it in the shim is
misleading.

Replace the use of the _fast types in our code with the fixed sized one,
it is more readable.

Change-Id: Id1f90ee374c0ca3838536ab6495c373ea1984fcf

rip-it: f1e2426bb25fb656e72c557f12ed13f7e5bc5ac0