iopy: rework void field handling.
Required void fields in structures do not bring any information (they are
used for removing a field in a backward-compatible manner), and they
should he handled as if they did not exist (except it should be possible
to set them with any value).

So in a iopy python object, we should *never* have such a field present.
The previous code was setting them to 'None'.

Change-Id: I1305b008dcd925fcc01cc3242a68e22d382365cd

rip-it: bfdc0d20afd42f9e97e7226d24d992d635161d23