yaml: improve handling of symbolic links when including.
When including or parsing a symbolic link, we want to resolve paths
relative to the real file:

toto:
  file.yml # contains an "!include:included.yml"
  included.yml
dir:
  file.yml -> ../toto/file.yml

When parsing dir/file.yml, the include of included.yml should work
properly. This was not done, as the file was searched in "dir" and not
in "toto".

To solve this, we must resolve any fullpath to an include into the
"real" path to the file. In addition, the "root" directory (the
protection that ensures all includes are contained in the directory
of the root file) must be updated, so that we can symlink outside of the
original root directory. Otherwise, the previous example would fail, as
included.yml is not inside "dir" which is the root directory.

Change-Id: Icbe2c8a0dfbac4f4ac4764fe6d70e56ec3f76b17

rip-it: dc455bd095049b339df0dca84b9bcdec5eb18afd