wscript: remove dependency to clang-format.
`clang-format` was not actually used as dependency, but it was used to
get the path to clang installation directory.
We cannot use `clang` directory as it be hidden behind ccache.

Fortunately, we can avoid this by using `-print-prog-name` argument of
clang that give us the true path of clang.

Also:
Use clang from CC variable if available instead of the default one.

Add a check that `clang-c/Index.h` is well installed to avoid
failing at compilation.

Change-Id: I25c1b2d4e3f171803e94649cc90c8f3d6c6454d0

rip-it: 961e0804651bb0f246f41d0d0a6ee7a0eeb6f719