iopc: make void parameters optional in typescript typings
Void parameters are mostly used for backward compat. It does not have to
be specified when creating an IOP type, but it was declared as mandatory
in the typings:

struct Toto {
    string name;
    void obsolete;
};

in typescript:

let toto: Toto_ModelParam {
    name: "aa",
    obsolete: null, // mandatory per TS typings
};

To avoid this, mark void parameters as optional even if required. It
does not matter, because a required void field is useless anyway.

However, keep it as mandatory if part of a union. This is because it can
be used as a discriminant:

union Toto {
    string version_a;
    void version_b;
};

in typescript:

type Toto = { name: string } | { version_b: null };

Change-Id: Iaecbea3c3b66f65c254245ec6867f4c8dfd83d9e

rip-it: b30ee70d3b50fb077f295650d2fe9dda71da743b