lib-priv[licence]: Fix test failing since today (august 9th 2020).
(cherry picked from mmsx commit a61116835369b8518b792f9210b42ceadcf97036)

The product expiration is encoded in activation tokens as a number of days
from the signature timestamp. And today, one of these tests have reached
its expiration date, computed from committed signature timestamp
1509098400.

To definitively fix this issue, just set a signature timestamp 10 years in
the future (which is accepted in this context).

Restore the signature timestamp 1509098400 for the end of the tests,
because they are made to work with it.

Change-Id: Ifeeace5fed0cf1ab9de3eaaf5a6b0edf4572ce9e

Conflicts:
	samples/licence-iop-2016.4-ok.cf
	zchk-licence.c

rip-it: 46c42111092e89999ed90526fcede1c729040380 uprooted