yaml: implement ser/deser for seq and obj yaml objects
When some yaml data is overridden and "lost" in the final document,
it is saved in the presentation, so that the document can be regenerated
properly when dumped.
This "save" is done by serializing the yaml data as IOP, but that was
only implemented for scalar values. This is now implemented for seq and
objs, so that it is possible to override any data.

Change-Id: Id5f573a5947136ea66a139c621040da3fc8799fa

rip-it: 90d48d6d228e1851b8a5d0fd557c9459687ed359