yaml: rework handling of variables
Variables were handled that way:
* when a variable is detected, add in the parsing env a map from the
  variable name to a pointer of the yaml_data_t object using the
  variable.
* When a variable definition is met, use the map to easily access all
  the yaml_data_t objects using the variable, and modify them.

This was optimal and prevented having to go through the AST to find all
the use of the variable. However, it was somewhat buggy, as the code
did copy sometimes yaml_data_t objects in-place, leading to the
"pointer" to the object no longer pointing to the right data (the one in
the AST). This caused a few bugs that were "avoided" by carefully
avoiding in-place copy. However, #73984 showed that this caused a bug
when using variables with merge keys.

Instead of continuing with workarounds, rework the handling of
variables. Now, we only track a set of which variables are
"unbounded", and go through the AST when replacing variables. This is
much worse performance-wise (even more so that we do this for every
variable), but much safer. And the YAML handling has never been designed
to put performance first at all, so it is fine.

Change-Id: Ie203afce6a4ad0940f6c7db1d5f6f1389ea11b53

rip-it: bbf861a3d27ed3ad93340ffadfddba147ce1655d