yaml: improve replacement of variables
When multiple variables are being set, go through the AST once and
iterate on the variables on every variable use, rather than going
through the AST for every variable binding.
This is much better performance-wise, as the most common use case is to
set many variables in a single place in the document.

Change-Id: I197ca96975d50259626730c871a10cc24d56f698

rip-it: d51ada04957398e9035e5ebc8476187c8e2b9dbd