yaml: forbid use of variables in vars used in templated string
Currently, it is possible to do something like this:

a: "$(a) $(b) $(c)"

...

variables:
  b: "$(d) \\$(e) $(c)"

This would slice some variables (and escaped variables) into the
templated string. This is dealable on parsing. However, it makes for
extremely complex code if we want to handle those on repacking (as the
AST might have changed, and we want to deduce the new values of all
variables).

So instead of allowing a broken feature, just forbid use of variables in
this context.

Change-Id: I0fbf988e1857b96fbebc3eb222093727c7dbdad2

rip-it: 652d119c7b743d6ef64d62900e4f229f6fa1e0ca