iop-yaml: fix conversion of json included subfiles with classes
See previous commit. When the iop field path was going through children
of base classes, the resolution of the type of the included field
failed, leading to a non-raw include being generated.
For example, with:

class Parent {};
class Child : Parent {
    string textFile;
};
class Foo {
    Parent p;
};

and a JSON for Foo such as:

{
    "p": {
        "_class": "Child",
        "textFile": @include("lorem_ipsum.txt")
    }
}

The include in the yaml version was not a raw include, leading to the
text file not being packed properly.

Change-Id: Ia5729fd7605efc1721700d98cc049cb26196442b

rip-it: 00437bb7e6124521bf58cb87fb75d19d22adda68