parseopt: expose int parser for positional arguments
This is a common issue with parseopt: if I want to have a positional
argument that is a string, then I just have to use NEXTARG or NEXTOPTARG.

If I want to have a positional argument that is an integer, then I have to
parse it by myself: I cannot rely on parseopt for that.

A common consequence is that people will begin using options for
integer parameters that are mandatory, making the command they are
creating a bit heavier to use.

Forged example:

 $ ./randrange --min 4 --max 10
 7

By exposing this kind of helpers, it will be easier to do that instead:

 $ ./randrange 4 10
 7

Note: I'm submitting it for 18.3 because I'd like to fix some TODOs in
18.3 tools and because this patch cannot introduce bugs in any way.

Change-Id: I9f82f6ed20b6bfd00094d3fdfd66abfb284336f7

rip-it: f88226d7ebb1a2099e3fc997ebb339509b30c321 uprooted