waf: Get rid of debug sections compression.
On recent versions of RedHat/gdb, the symbols compression make it
impossible to debug binaries properly.
We have opened this ticket, but it was rejected:
https://bugzilla.redhat.com/show_bug.cgi?id=1877286

So we have decided to simply remove the compression, because this is a
source of problems and does not really bring strong advantages.
This will make the debug binaries bigger, but make the compilation faster
and won't impact the size of release binaries.

Change-Id: I4f9911218368eb841b2a414f203497713267e040

rip-it: 593bba79cd80a0c900b49cd833da56b23bcba4b1 uprooted