make: Get rid of debug sections compression.
On recent versions of RedHat/gdb, the symbols compression make it
impossible to debug binaries properly.
We have opened this ticket, but it was rejected:
https://bugzilla.redhat.com/show_bug.cgi?id=1877286

So we have decided to simply remove the compression, because this is a
source of problems and does not really bring strong advantages.
This will make the debug binaries bigger, but make the compilation faster
and won't impact the size of release binaries.

Change-Id: I5343d679fcbf49f50b15d0fd30af98804f227a07

Conflicts:
	Build/README
	Build/bulk-library.mk
	Config/cflags.mk

rip-it: d7893f6bbfe3a1ba06c21625be60a8b59877cbb7 uprooted