yaml: fix pack of empty array/objects after override
When packing a data that was modified from an override, the overridden
fields are skipped, as they will be packed when the override itself
is packed.
However, if all fields come from an override, this lead to no packing
at all, which in yaml is equivalent to the "nil" value. This lead to
an error when reading the packed yaml data, as overriding the nil value
was invalid.

This is fixed by checking after the packing of an object or array if
anything was written. If not, the empty object/array is written.

Change-Id: I26fe16602d02c0746b23ef7f2df16da3fc9990dd

rip-it: e13d9ee89bc32b29d16d2fe73a085f39542b2e7a