yaml: fix handling of presentation data with includes
= Description of the bug =

When reformatting includes, the result was sometimes inserting newlines
at wrong positions, for example:

a:
  - 2
  - !include:foo.yml

b: 3

could result in

a:
  - 2
  -
    !include:foo.yml
b: 3

To explain what happens, we have to describe two features first.

= YAML null values =

Lets have fun with YAML null values! In YAML, there is a great feature
where the absence of data is not invalid, but an indication of a null
value.
For example, this:

a: 1
b: ~
c: 2

is the same as:

a: 1
b:
c: 2

The key 'b' has the value 'null'. This is stupid, but well, we are
talking about YAML.

This feature is used for includes, where the syntax is this one:

a: !include:file.yml
b: 2

This is actually parsed as a tagged null value:

a: !include:file.yml ~
b: 2

(For why we use this syntax, see d6ab8cb7830591).

= Prefix comments presentation =

In our YAML implementation, we want to save presentation data, so that
when repacked, the file is dumped without any changes if possible.
For this reason, we associate with YAML data several presentation data,
including:
* comments (prefix, and inline)
* prefix newlines

For example:

a: 1

b: 2

has a presentation data that ".b" has 1 prefix newline.

As YAML is hell to parse and requires severe lookahead, the prefix
comments/newlines is hard to handle correctly. What happens is that we
accumulate prefix data while we lookahead, and when we finally create a
new YAML data, we add accumulated prefix data to it.
In the previous example, we would parse "1", then search ahead,
accumulate 1 prefix newline, and when parsing "b", add the prefix
newline to the YAML data.

This works fine... except when you use the "no data is null data"
feature as well.

This example:

a:

b: 2

would lead to this behavior:

* parse "a:"
* lookahead to search for some data, accumulate 1 prefix newline
* find that there is no data, so create a "null data".
* this data creation will use the accumulated prefix data, so the null
  data now has 1 prefix newline.

leading to this repack:

a:

b: 2

well nothing changed here, but with a tag for the null data:

a:
  !tag
b: 2

and here's the bug.

= The fix =

the fix is twofold

* when creating the null data from no data, do not use accumulated
  prefix presentation.
* we need to fix the parsing of a tag a bit, so that comments inserted
  before the tag and the value do not break the presentation. The tag
  is now directly added to a new data, before parsing the rest of the
  data. This requires the parse_data function to work on an already
  initialized out value, and the start_data function to be reentrant.

Change-Id: Icb7c575a4f0ced6fba7fdddff3e793ff9d5bde31

rip-it: ddb34b8b7c827698aa6663a91d01532f7de824e3