z: fix broken escaping for string comparisons
When printing strings for after a string comparison failed, the Z library
tries to escape some special characters and fallback to hexadecimal if
there is still a non-printable character in the string to display.

Problem: the escaping is useless because it's done *after* detecting
non-printable characters. Consequence: we pointlessly only escape the
backslash '\' and switch to hexadecimal as soon as we meet a '\n'.

Solution: just look for printable characters *after* escaping. It could be
faster but who cares ? We only use the function when a test fails.

Change-Id: I0389464770f9a3d1e6aec0bcc4df379e96d808d9

rip-it: 6e39049e15b0c63e0f63c9be91f36ba5437a8c19 uprooted