ssl: Add precious warnings in the documentation of ssl_write.
The documentation was pretending that this function mimics write(2), but
it actually differs in two major aspects. Warn about it.

Note that the second point *really* sucks, and makes it quite hard to
adapt some of our protocols to SSL...

Change-Id: I558ddf0b34a6eaf0648d1879061d692132bd3fb7

rip-it: 7a6d935e8c99d3c2eb6bd069259da4a0fe2fa686