core_dump: adapt find_binary_fullpath to gdb output
On CentOS 6 the output was like:
"Core was generated by `test-cd -n test_arg'"
but on CentOS 7 we now have the full path:
"Core was generated by `/full/path/test-cd -n test_arg'"

So the "first check" was always failing because we were checking the
binary in the wrong path.

Note that it does not make sens to split the regex answer as we don't
match the spaces in the regex.

Change-Id: I8647b51c4effe38611c6ddd637c2bf059b45fce2

rip-it: 6644b3c52e08e72ed7b7d58f13879e6d69519934