httpc: Inline TLS certificates in IOP configurations.
The tlsCertPath IOP field, introduced by cd05ffc7799, was the path to the
TLS certificate file, and not the content of the certificate.

As this structure is meant to be used in our products configurations, this
choice was causing several issues:
* the certificates have to be manually synchronized on all the hosts using
  them, independently from the configuration (that is natively
  replicated), which is annoying and error-prone.
* when using relative path, the code could fail depending on the current
  directory of the program, which is fragile.
* using absolute paths (which solves the previous issue) is not suited for
  committed development configurations.

So change the way it works, and inline the certificate instead (which can
easily be done using the json/yaml include feature). Luckily, this field
is not used in production yet, so we can safely break the backward
compatibility.

Change-Id: Ib1a7ac4fbf555a25409d75be407692956e1c94e6

rip-it: 9cce775c47ce6b7e5941e7e330de4d7ed089c33c