openapi: Skip abstract classes when listing possible children in oneOf.
"oneOf" is used to list the possible types of a field. So when a field has
a class type, "oneOf" must contains the actual class of the field, and
its children classes.

But the abstract classes must be excluded from this list, as they cannot
be used as the real type of the field. So skip them, just like it was done
for private classes.

Change-Id: Ic6409a176beab5638482f480e70c5d3c168be6e6

rip-it: 7d91d3400a6965d9ec8f335f84c24b7dba3fc2d0