yaml.iop: Fix documentation for doxygen syntax.
The previous documentation of the PresentationNode::emptyLines field was
triggering the following antora warning when generating the
documentation:

  section title out of sequence: expected level 1, got level 2

This is because "---" was wrongly interpreted as a section.

Fix this by using "-----", and use it everywhere else to have a great
rendering.

Change-Id: Ic870661f2702496afa7530fda7180a16969e4ac9

rip-it: 07e4df0a608c3416110f694ac399fe6c5899d067