iopc/lang-c: stop changing the order of vector iopc_struct_t::fields
In iopc-lang-c.blk, the fields of a struct could be sorted in three
ways:

1. In the original order of declaration (by pos);
2. Sorted by field tag;
3. Ordered for C struct declaration in a way that minimize the struct
   size.

Each time iopc-lang-c needed the fields to be sorted in a particular
order, the vector 'fields' was modified and the current order saved in
the attribute 'fields_sort_order'.

This way to do is eccentric but it became unacceptable when the sorting
functions was exported so other languages implementations (JSON and
typescript) could put the fields back in the original order.

This patch gets rid of this "curiosity" by adding two extra vectors in
iopc_struct_t, so each order has its dedicated vector and the vector
'fields' has a consistent order.

Change-Id: Ie4563226db25546fd1eeda69238d5d953cced5df

rip-it: f688762192652cd01ff5af7fde4c3077f80db5fe