[BACKPORT] waf/pylint: fix pylint command with waf.
We should always run the pylint command from the project root directory,
and not from the current directory.
This is because we need to run the pylint command where the `pylintrc`
file is.
So we should use `cwd` to set the current directory when running the
command.

Change-Id: I952f6785fd75f6991ff460573948a23aa118f522
(cherry picked from commit 63ab28b7349170d202455b32ea1c90d129e6c8bf)
Priv-Id: d1130790e0ca68d1c9b5176f04bff6485cff9dff