waf/pylint: fix pylint command with waf.
We should always run the pylint command from the project root directory,
and not from the current directory.
This is because we need to run the pylint command where the `pylintrc`
file is.
So we should use `cwd` to set the current directory when running the
command.

Change-Id: I952f6785fd75f6991ff460573948a23aa118f522

rip-it: 63ab28b7349170d202455b32ea1c90d129e6c8bf