pxcc: use clang flags for PXCC and ignore unused variables.
When parsing header files with PXCC, we use Clang.
There are some warnings that we ignore that are set by default with
Clang.
So we should ignore these warning as well with PXCC.

Consecutively, we now need to ignore the unused variables created by
PXCC on parsing.

Change-Id: I17bfa9e30a8ebc5fa8c84496ac84e946617e9c08

rip-it: 41b2158dd569d6f2e14ef20a8111851e2ba8e49d