cython: rework lib-common cython source files directory.
Instead of having a `src/cython`, we now have a `src/libcommon_cython`.
We now include `src` instead of `src/cython`.
This way we have a nice package `libcommon_cython`, and instead of
doing this:

    from libcommon_core cimport *

We now have this:

    from libcommon_cython.core cimport *

Change-Id: I4580ef860970a427646889eab5719dac27c5e19b

rip-it: 782d547e8dd9af447ca343aa296df6cb75c15d76